+++ sertit/vectors.py
@@ -188,14 +188,17 @@ def set_kml_driver() -> None:
         [1 rows x 12 columns]
 
     """
-    import fiona
+    from importlib.metadata import version
 
-    drivers = fiona.drvsupport.supported_drivers
+    if int(version("geopandas").split(".")[0]) <= 1.0:
+        import fiona
 
-    if "LIBKML" not in drivers:
-        drivers["LIBKML"] = "rw"
-    if "KML" not in drivers:  # Just in case
-        drivers["KML"] = "rw"
+        drivers = fiona.drvsupport.supported_drivers
+
+        if "LIBKML" not in drivers:
+            drivers["LIBKML"] = "rw"
+        if "KML" not in drivers:  # Just in case
+            drivers["KML"] = "rw"
 
 
 def get_aoi_wkt(aoi_path: AnyPathStrType, as_str: bool = True) -> Union[str, Polygon]:
@@ -576,13 +579,24 @@ def _read_kml(
     # When you try to get the KML content, you actually get the first layer.
     # So you need for loop for iterating over layers.
     # https://gis.stackexchange.com/questions/328525/geopandas-read-file-only-reading-first-part-of-kml/328554
+    from importlib.metadata import version
+
     import fiona
+    from pyogrio.errors import DataSourceError
 
     driver = "KML" if gpd_vect_path.endswith(".kml") else "KMZ"
+
+    # WORKAROUND: https://github.com/geopandas/pyogrio/issues/444
+    if version("geopandas") >= "1.0.0" and version("pyogrio") <= "0.10.0":
+        engine = "fiona"
+    else:
+        engine = None
+
     for layer in fiona.listlayers(gpd_vect_path):
         try:
+
             vect_layer = gpd.read_file(
-                gpd_vect_path, driver=driver, layer=layer, **kwargs
+                gpd_vect_path, driver=driver, layer=layer, engine=engine, **kwargs
             )
             if not vect_layer.empty:
                 # KML files are always in WGS84 (and does not contain this information)
@@ -590,6 +604,10 @@ def _read_kml(
                 vect = pd.concat([vect, vect_layer])
         except ValueError:
             pass  # Except Null Layer
+        except DataSourceError:
+            LOGGER.error(
+                f"Error in reading {path.get_filename(gpd_vect_path)}. geopandas: {version('geopandas')},  pyogrio: {version('pyogrio')}, engine: {engine}"
+            )
 
     # Workaround for archived KML -> they may be empty
     # Convert KML to GeoJSON