+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 1.37.3 (2024-mm-dd)
 
 - FIX: Manage the duality of exceptions between `fiona` and `pyogrio` for `DriverError`/`DataSourceError`
+- FIX: Handle null GMLs with `pyogrio` engine
 - FIX: Workaround for KMZ issue with `geopandas > 1.0` and `pyogrio < 0.10.0` ([Github issue](https://github.com/geopandas/pyogrio/issues/444))
 
 ## 1.37.2 (2024-05-22)